Module xtask::cargo_commands

Functions

build_and_check_size

Check the binary sizes of examples

cargo

Cargo command to either build or check

cargo_book

Use mdbook to build the book

cargo_clippy

Run cargo clippy on selected package

cargo_doc

Run cargo doc

cargo_example

Cargo command to either build or check all examples

cargo_format

Run cargo fmt on selected package

cargo_test

Run cargo test on the selcted package or all packages

run_test

Run examples