Module xtask::codegen

We use code generation heavily in rust-analyzer.

Rather then doing it via proc-macros, we use old-school way of just dumping the source code.

This module’s submodules define specific bits that we generate.

Modules

gen_assists_docs
Generates assists.md documentation.

gen_diagnostic_docs
Generates assists.md documentation.

gen_feature_docs
Generates assists.md documentation.

gen_lint_completions
Generates descriptors structure for unstable feature from Unstable Book

gen_parser_tests
This module greps parser’s code for specially formatted comments and turnes them into tests.

gen_syntax
This module generates AST datatype used by rust-analyzer.

Structs

CommentBlock

Location

Constants

PREAMBLE

Functions

do_extract_comment_blocks

docs

ensure_file_contents
Checks that the file has the specified contents. If that is not the case, updates the file and then fails the test.

extract_comment_blocks

extract_comment_blocks_with_empty_lines

normalize_newlines

reformat

used