xtask::commands

Module rextendr_document

This invokes rextendr::document() within tests/extendrtests.

It uses the vendored rextendr in the repository as the source package.

Ensure that git submodule update --init was invoked once, as to setup the vendored rextendr package.

devtools must be installed on system.

The idea here is to be able to develop rextendr alongside extendr, as well as ease the development of extendr.

Functions

run 🔒

run_rextendr_document 🔒