Module xtask::dump::build

Constants

BRANCH

display current branch

BUILD_OS

display build system os

BUILD_RUST_CHANNEL

display project build by rust channel [debug or release]

BUILD_TARGET

display build current project target

BUILD_TARGET_ARCH

display build current project version arch

BUILD_TIME

display project build time

BUILD_TIME_2822

display project build time by rfc2822

BUILD_TIME_3339

display project build time by rfc3399

CARGO_TREE

display build cargo dependencies.It’s used by rust version 1.44.0

CARGO_VERSION

display build system cargo version

CLAP_LONG_VERSION

The common version const. It’s so easy to use this const with CLAP_VERSION.

CLAP_VERSIONDeprecated

The common version const. It’s so easy to use this const with CLAP_VERSION.

COMMIT_AUTHOR

display current commit author

COMMIT_DATE

display current commit date

COMMIT_DATE_2822

display current commit date by rfc2822

COMMIT_DATE_3339

display current commit date by rfc3339

COMMIT_EMAIL

display current commit email

COMMIT_HASH

display current commit_id

GIT_CLEAN

display current git repository status clean:‘true or false’

GIT_STATUS_FILE

display current git repository status files:‘dirty or stage’

PKG_DESCRIPTION

display build current project description

PKG_VERSION

display build current project version

PKG_VERSION_MAJOR

display build current project major version

PKG_VERSION_MINOR

display build current project minor version

PKG_VERSION_PATCH

display build current project patch version

PKG_VERSION_PRE

display build current project preview version

PROJECT_NAME

display project name

RUST_CHANNEL

display build system rust channel

RUST_VERSION

display build system rust version

SHORT_COMMIT

display current short commit_id

TAG

display current tag

VERSION

The common version const. It’s so easy to use this const.

Functions

print_build_in

print build in method