xtask

Enum Commands

pub(crate) enum Commands {
    RunTests(TestArgs),
    Clippy(ClippyArgs),
    Tidy(TidyArgs),
    SemverChecks(SemverArgs),
}

Variants

RunTests(TestArgs)

Run all tests for diesel

Requires cargo-nextest to be installed

Clippy(ClippyArgs)

Run clippy on all crates

Tidy(TidyArgs)

Perform a set of preliminary checks

This command will execute cargo fmt --check to verify that the code is formatted, typos to check for spelling errors and it will execute xtask clippy to verify that the code compiles without warning

SemverChecks(SemverArgs)

Check for semver relevant changes

This command will execute cargo semver-checks to verify that no breaking changes are included

Implementations

impl Commands

pub(crate) fn run(self)