Module xtask::examples

Structs

Executable 🔒

OakExample

Holds the components for running an example in Oak: a server running the given applications listening to the clients and passing requests to backends. Configured through example.toml files.

OakFunctionsApplication 🔒

OakFunctionsExample

Server 🔒

WizerOpt 🔒

Enums

Application 🔒

Identify the application the server runs.

Target

Constants

ALL_CLIENTS

DEFAULT_EXAMPLE_BACKEND_RUST_TARGET 🔒

DEFAULT_SERVER_RUST_TARGET 🔒

NO_CLIENTS

Traits

OakExampleSteps 🔒

Construct run and build steps an OakExample.

Functions

build 🔒

build_docker 🔒

build_oak_functions_example

build_oak_functions_server_variants

Build every variant of the function server. It’s easier to always build all variants than to keep track of which variant to build and the overhead of building all variants is acceptable.

build_rust_binary 🔒

build_wasm_module

run 🔒

run_client 🔒

run_clients 🔒

run_oak_functions_example 🔒

run_oak_functions_examples

run_oak_functions_server 🔒

run_wizer 🔒