Module xtask::files

Functions

crate_manifest_files

Return an iterator of all known Cargo Manifest files that define crates.

file_contains

is_bazel_file

Return whether the provided path refers to a Bazel file (BUILD, WORKSPACE, or *.bzl)

is_build_file

is_clang_format_file

Return whether the provided path refers to a source file that can be formatted by clang-tidy.

is_dockerfile

is_fuzzing_toml_file

Return whether the provided path refers to a fuzz crate for fuzz testing with cargo-fuzz.

is_html_file

is_javascript_file

is_markdown_file

Return whether the provided path refers to a markdown file (*.md)

is_shell_script

is_source_code_file

Return whether the provided path refers to a source file in a programming language.

is_toml_file

is_typescript_file

is_yaml_file

read_file

source_files

Return an iterator of all the first-party and non-ignored files in the repository, which can be then additionally filtered by the caller.

to_string

workspace_manifest_files

Return an iterator of all known Cargo Manifest files that define workspaces.