Struct xtask::internal::Context

pub struct Context {
    opt: Opt,
    prefix: Vec<String>,
}

Encapsulates all the local state relative to a step, and is propagated to child steps.

Fields
opt: Optprefix: Vec<String>
Implementations

impl Context

pub fn root(opt: &Opt) -> Self

fn child(&self, name: &str) -> Self

fn depth(&self) -> usize

fn header(&self) -> String

fn footer_long(&self) -> String

Prints a footer that repeats information from the header.

Useful when footer and header are expected to be far away from each other.

fn footer_short(&self) -> String

Prints a footer that does not repeat information from the header.

Useful when it follows the header almost immediately.

fn margin(&self) -> String