Module xtask::launcher

Functionality for testing variants of the enclave binary exposed by the launcher.

Structs

App

Statics

MOCK_LOOKUP_DATA_PATH

OAK_RESTRICTED_KERNEL_WRAPPER_BIN

Functions

build_binary

build_stage0

just_build

run_launcher

run_oak_functions_example_in_background

Runs the specified example as a background task. Returns a reference to the running server and the port on which the server is listening.

run_oak_functions_launcher_example

Runs the Oak Functions launcher configured with a default Wasm module for key / value lookups and mock lookup data.

run_oak_functions_launcher_example_with_lookup_data