xtask

Module pgo

PGO (Profile-Guided Optimization) utilities.

Functions

apply_pgo_to_cmd 🔒

build_command 🔒

Helper function to create a build command for rust-analyzer

download_crate_for_training 🔒

Downloads a crate from GitHub, stores it into pgo_dir and returns a path to it.

gather_pgo_profile 🔒

Decorates ra_build_cmd to add PGO instrumentation, and then runs the PGO instrumented Rust Analyzer on itself to gather a PGO profile.