Trait xtask::runtime::runtime_v1::RuntimeV1

pub trait RuntimeV1: Sized + WasmerEnv + 'static {
    type ArgumentHint: Debug;
    type ArgumentMetadata: Debug;
    type GraphContext: Debug;
    type KernelContext: Debug;
    type Metadata: Debug;
    type Model: Debug;
    type TensorHint: Debug;
    type TensorMetadata: Debug;

    fn metadata_new(&mut self, name: &str, version: &str) -> Self::Metadata;
    fn metadata_set_description(
        &mut self, 
        self_: &Self::Metadata, 
        description: &str
    );
    fn metadata_set_repository(&mut self, self_: &Self::Metadata, url: &str);
    fn metadata_set_homepage(&mut self, self_: &Self::Metadata, url: &str);
    fn metadata_add_tag(&mut self, self_: &Self::Metadata, tag: &str);
    fn metadata_add_argument(
        &mut self, 
        self_: &Self::Metadata, 
        arg: &Self::ArgumentMetadata
    );
    fn metadata_add_input(
        &mut self, 
        self_: &Self::Metadata, 
        metadata: &Self::TensorMetadata
    );
    fn metadata_add_output(
        &mut self, 
        self_: &Self::Metadata, 
        metadata: &Self::TensorMetadata
    );
    fn argument_metadata_new(&mut self, name: &str) -> Self::ArgumentMetadata;
    fn argument_metadata_set_description(
        &mut self, 
        self_: &Self::ArgumentMetadata, 
        description: &str
    );
    fn argument_metadata_set_default_value(
        &mut self, 
        self_: &Self::ArgumentMetadata, 
        default_value: &str
    );
    fn argument_metadata_add_hint(
        &mut self, 
        self_: &Self::ArgumentMetadata, 
        hint: &Self::ArgumentHint
    );
    fn tensor_metadata_new(&mut self, name: &str) -> Self::TensorMetadata;
    fn tensor_metadata_set_description(
        &mut self, 
        self_: &Self::TensorMetadata, 
        description: &str
    );
    fn tensor_metadata_add_hint(
        &mut self, 
        self_: &Self::TensorMetadata, 
        hint: &Self::TensorHint
    );
    fn interpret_as_image(&mut self) -> Self::TensorHint;
    fn interpret_as_audio(&mut self) -> Self::TensorHint;
    fn supported_shapes(
        &mut self, 
        supported_element_types: Vec<ElementType>, 
        dimensions: DimensionsParam<'_>
    ) -> Self::TensorHint;
    fn interpret_as_number_in_range(
        &mut self, 
        min: &str, 
        max: &str
    ) -> Self::ArgumentHint;
    fn interpret_as_string_in_enum(
        &mut self, 
        string_enum: Vec<&str>
    ) -> Self::ArgumentHint;
    fn non_negative_number(&mut self) -> Self::ArgumentHint;
    fn supported_argument_type(
        &mut self, 
        hint: ArgumentType
    ) -> Self::ArgumentHint;
    fn register_node(&mut self, metadata: &Self::Metadata);
    fn graph_context_for_node(
        &mut self, 
        node_id: &str
    ) -> Option<Self::GraphContext>;
    fn graph_context_get_argument(
        &mut self, 
        self_: &Self::GraphContext, 
        name: &str
    ) -> Option<String>;
    fn graph_context_add_input_tensor(
        &mut self, 
        self_: &Self::GraphContext, 
        name: &str, 
        element_type: ElementType, 
        dimensions: DimensionsParam<'_>
    );
    fn graph_context_add_output_tensor(
        &mut self, 
        self_: &Self::GraphContext, 
        name: &str, 
        element_type: ElementType, 
        dimensions: DimensionsParam<'_>
    );
    fn kernel_context_for_node(
        &mut self, 
        node_id: &str
    ) -> Option<Self::KernelContext>;
    fn kernel_context_get_argument(
        &mut self, 
        self_: &Self::KernelContext, 
        name: &str
    ) -> Option<String>;
    fn kernel_context_get_input_tensor(
        &mut self, 
        self_: &Self::KernelContext, 
        name: &str
    ) -> Option<TensorResult>;
    fn kernel_context_set_output_tensor(
        &mut self, 
        self_: &Self::KernelContext, 
        name: &str, 
        tensor: TensorParam<'_>
    );
    fn kernel_context_get_global_input(
        &mut self, 
        self_: &Self::KernelContext, 
        name: &str
    ) -> Option<TensorResult>;
    fn kernel_context_set_global_output(
        &mut self, 
        self_: &Self::KernelContext, 
        name: &str, 
        tensor: TensorParam<'_>
    );
    fn is_enabled(&mut self, metadata: LogMetadata<'_>) -> bool;
    fn log(
        &mut self, 
        metadata: LogMetadata<'_>, 
        message: &str, 
        data: LogValueMap<'_>
    );
    fn model_load(
        &mut self, 
        model_format: &str, 
        model: &[u8], 
        arguments: Vec<(&str, &str)>
    ) -> Result<Self::Model, ModelLoadError>;
    fn model_infer(
        &mut self, 
        self_: &Self::Model, 
        inputs: Vec<TensorParam<'_>>
    ) -> Result<Vec<TensorResult>, ModelInferError>;
    fn model_inputs(&mut self, self_: &Self::Model) -> Vec<Shape>;
    fn model_outputs(&mut self, self_: &Self::Model) -> Vec<Shape>;

    fn drop_argument_hint(&mut self, state: Self::ArgumentHint) { ... }
    fn drop_argument_metadata(&mut self, state: Self::ArgumentMetadata) { ... }
    fn drop_graph_context(&mut self, state: Self::GraphContext) { ... }
    fn drop_kernel_context(&mut self, state: Self::KernelContext) { ... }
    fn drop_metadata(&mut self, state: Self::Metadata) { ... }
    fn drop_model(&mut self, state: Self::Model) { ... }
    fn drop_tensor_hint(&mut self, state: Self::TensorHint) { ... }
    fn drop_tensor_metadata(&mut self, state: Self::TensorMetadata) { ... }

}

Required Associated Types

type ArgumentHint: Debug

type ArgumentMetadata: Debug

type GraphContext: Debug

type KernelContext: Debug

type Metadata: Debug

type Model: Debug

type TensorHint: Debug

type TensorMetadata: Debug

Required Methods

fn metadata_new(&mut self, name: &str, version: &str) -> Self::Metadata

Create a new metadata object with the provided name and version number.

The name should typically be one or two words that concisely describe the node and will be used as the human-friendly label shown to users when referring to it.

fn metadata_set_description(
&mut self,
self_: &Self::Metadata,
description: &str
)

A human-friendly description of the node.

The text may include markdown.

fn metadata_set_repository(&mut self, self_: &Self::Metadata, url: &str)

A repository containing this node’s source code.

fn metadata_set_homepage(&mut self, self_: &Self::Metadata, url: &str)

The node’s home page.

This will typically point to a README file or a page on the internet that users can go to when they want to find out more about the node.

fn metadata_add_tag(&mut self, self_: &Self::Metadata, tag: &str)

Associate this node with a particular tag.

Tags are typically used to assist in search and filtering.

fn metadata_add_argument(
&mut self,
self_: &Self::Metadata,
arg: &Self::ArgumentMetadata
)

Arguments this node accepts.

fn metadata_add_input(
&mut self,
self_: &Self::Metadata,
metadata: &Self::TensorMetadata
)

Information about this node’s input tensors.

fn metadata_add_output(
&mut self,
self_: &Self::Metadata,
metadata: &Self::TensorMetadata
)

Information about this node’s output tensors.

fn argument_metadata_new(&mut self, name: &str) -> Self::ArgumentMetadata

Create a new named argument.

fn argument_metadata_set_description(
&mut self,
self_: &Self::ArgumentMetadata,
description: &str
)

A human-friendly description of the argument.

The text may include markdown.

fn argument_metadata_set_default_value(
&mut self,
self_: &Self::ArgumentMetadata,
default_value: &str
)

A useful default value for this argument.

fn argument_metadata_add_hint(
&mut self,
self_: &Self::ArgumentMetadata,
hint: &Self::ArgumentHint
)

fn tensor_metadata_new(&mut self, name: &str) -> Self::TensorMetadata

Create a new named tensor.

fn tensor_metadata_set_description(
&mut self,
self_: &Self::TensorMetadata,
description: &str
)

A human-friendly description of the tensor.

The text may include markdown.

fn tensor_metadata_add_hint(
&mut self,
self_: &Self::TensorMetadata,
hint: &Self::TensorHint
)

Add a hint that provides the runtime with contextual information about this node.

fn interpret_as_image(&mut self) -> Self::TensorHint

Hint to the runtime that a tensor may be displayed as an image.

fn interpret_as_audio(&mut self) -> Self::TensorHint

Hint to the runtime that a tensor may be interpreted as an audio clip.

fn supported_shapes(
&mut self,
supported_element_types: Vec<ElementType>,
dimensions: DimensionsParam<'_>
) -> Self::TensorHint

Hint that a tensor may have a particular shape and the element types it supports.

Note: This hint will be removed in the future in favour of a more flexible mechanism.

fn interpret_as_number_in_range(
&mut self,
min: &str,
max: &str
) -> Self::ArgumentHint

Hint to the runtime that an argument may be interpreted as a number in [min, max]

fn interpret_as_string_in_enum(
&mut self,
string_enum: Vec<&str>
) -> Self::ArgumentHint

Hint to the runtime that an argument may be interpreted as a string in a defined list

fn non_negative_number(&mut self) -> Self::ArgumentHint

Hint to the runtime that an argument may be interpreted as a non-negative number

fn supported_argument_type(&mut self, hint: ArgumentType) -> Self::ArgumentHint

Tell the runtime that this argument may have a certain type.

fn register_node(&mut self, metadata: &Self::Metadata)

Register a node type with the runtime.

fn graph_context_for_node(
&mut self,
node_id: &str
) -> Option<Self::GraphContext>

Get the graph context for a particular node.

Note: this should only be used from within the graph() function.

fn graph_context_get_argument(
&mut self,
self_: &Self::GraphContext,
name: &str
) -> Option<String>

fn graph_context_add_input_tensor(
&mut self,
self_: &Self::GraphContext,
name: &str,
element_type: ElementType,
dimensions: DimensionsParam<'_>
)

fn graph_context_add_output_tensor(
&mut self,
self_: &Self::GraphContext,
name: &str,
element_type: ElementType,
dimensions: DimensionsParam<'_>
)

fn kernel_context_for_node(
&mut self,
node_id: &str
) -> Option<Self::KernelContext>

Get the kernel context for a particular node.

Note: this should only be used from within the kernel() function.

fn kernel_context_get_argument(
&mut self,
self_: &Self::KernelContext,
name: &str
) -> Option<String>

Get a named argument.

fn kernel_context_get_input_tensor(
&mut self,
self_: &Self::KernelContext,
name: &str
) -> Option<TensorResult>

Get one of this node’s input tensors by name.

fn kernel_context_set_output_tensor(
&mut self,
self_: &Self::KernelContext,
name: &str,
tensor: TensorParam<'_>
)

Set one of this node’s input tensors by name.

fn kernel_context_get_global_input(
&mut self,
self_: &Self::KernelContext,
name: &str
) -> Option<TensorResult>

Get a global input tensor.

Note: these tensors live in a separate namespace from a node’s normal input and output tensors.

fn kernel_context_set_global_output(
&mut self,
self_: &Self::KernelContext,
name: &str,
tensor: TensorParam<'_>
)

Set a global input tensor.

Note: these tensors live in a separate namespace from a node’s normal input and output tensors.

fn is_enabled(&mut self, metadata: LogMetadata<'_>) -> bool

Check whether a particular message would be logged, allowing the guest to avoid potentially expensive work.

fn log(
&mut self,
metadata: LogMetadata<'_>,
message: &str,
data: LogValueMap<'_>
)

Record a log message with some structured data.

fn model_load(
&mut self,
model_format: &str,
model: &[u8],
arguments: Vec<(&str, &str)>
) -> Result<Self::Model, ModelLoadError>

Use the runtime to load a model.

The model-format argument tells the runtime what type of model is being loaded. Arguments, a list of key-value pairs, may be used to modify the loading process (e.g. to request hardware acceleration).

fn model_infer(
&mut self,
self_: &Self::Model,
inputs: Vec<TensorParam<'_>>
) -> Result<Vec<TensorResult>, ModelInferError>

fn model_inputs(&mut self, self_: &Self::Model) -> Vec<Shape>

The shape of this model’s input tensors.

fn model_outputs(&mut self, self_: &Self::Model) -> Vec<Shape>

The shape of this model’s output tensors.

Provided Methods

fn drop_argument_hint(&mut self, state: Self::ArgumentHint)

fn drop_argument_metadata(&mut self, state: Self::ArgumentMetadata)

fn drop_graph_context(&mut self, state: Self::GraphContext)

fn drop_kernel_context(&mut self, state: Self::KernelContext)

fn drop_metadata(&mut self, state: Self::Metadata)

fn drop_model(&mut self, state: Self::Model)

fn drop_tensor_hint(&mut self, state: Self::TensorHint)

fn drop_tensor_metadata(&mut self, state: Self::TensorMetadata)