Module xtask::testing

Structs

BackgroundStep

Thin wrapper around an inner Running that kills the inner Running when dropped.

Functions

run_background

Runs a step in the background, and returns a reference to the running process.

run_step

Runs a step, and asserts that it succeeds.

skip_test

Whether to skip the test. For instance, GitHub Actions does not support KVM, so we cannot run tests that require nested virtualization.