Trait xtask::traits::Benchmark

pub trait Benchmark {
    const ITERATIONS_PER_BENCHMARK: u64 = 3u64;

    fn benchmark(
        &self,
        metrics: &mut Metrics,
        name: &str,
        desc: &str
    ) -> Result<()>;
}

Trait Extension for the xshell::Cmd struct.

Provided Associated Constants

const ITERATIONS_PER_BENCHMARK: u64 = 3u64

Required Methods

fn benchmark(&self, metrics: &mut Metrics, name: &str, desc: &str) -> Result<()>

Benchmarks a command specified by the self argument measuring its wall-time, collecting and inserting the data into the Metrics struct. The name thereby specifies the to be benchmarked task (e.g. plc) whereas the desc argument describes how the task ran (e.g. with the -Oaggressive flag).