Crate xtask

See https://github.com/matklad/cargo-xtask/.

This binary defines various auxiliary build commands, which are not expressible with just cargo. Notably, it provides tests via cargo test -p xtask for code generation and cargo xtask install for installation of rust-analyzer server and client.

This binary is integrated into the cargo command line by using an alias in .cargo/config.

Modules

codegen ğŸ”’

dist ğŸ”’

flags ğŸ”’

install ğŸ”’

Installs rust-analyzer language server and/or editor plugin.

metrics ğŸ”’

publish ğŸ”’

release ğŸ”’

tidy ğŸ”’

util ğŸ”’

Functions

date_iso ğŸ”’

is_release_tag ğŸ”’

main ğŸ”’

project_root ğŸ”’

Returns the path to the root directory of rust-analyzer project.

run_fuzzer ğŸ”’