Crate xtask

See https://github.com/matklad/cargo-xtask/.

This binary defines various auxiliary build commands, which are not expressible with just cargo. Notably, it provides tests via cargo test -p xtask for code generation and cargo xtask install for installation of rust-analyzer server and client.

This binary is integrated into the cargo command line by using an alias in .cargo/config.

Modules

ast_src
Defines input for code generation process.

codegen
We use code generation heavily in rust-analyzer.

dist

flags

install
Installs rust-analyzer language server and/or editor plugin.

metrics

release

Functions

date_iso

ensure_rustfmt

files_in

is_release_tag

main

project_root

run_fuzzer

rust_files

rust_files_in