Module xtask::codegen::lints

Generates descriptor structures for unstable features from the unstable book and lints from rustc, rustdoc, and clippy.

Structs

ClippyLint ğŸ”’

Constants

DESTINATION ğŸ”’

Functions

find_and_slice ğŸ”’

generate ğŸ”’

This clones rustc repo, and so is not worth to keep up-to-date on a constant basis.

generate_descriptor_clippy ğŸ”’

generate_feature_descriptor ğŸ”’

Parses the unstable book src_dir and prints a constant with the list of unstable features into buf.

generate_lint_descriptor ğŸ”’

Parses the output of rustdoc -Whelp and prints Lint and LintGroup constants into buf.

push_lint_completion ğŸ”’

push_lint_group ğŸ”’

unescape ğŸ”’