Module xtask::internal

Structs

BuildClient

BuildServerOpt

CargoBinary

Partial information about a Cargo binary, as included in a Cargo manifest.

CargoManifest

Partial representation of Cargo manifest files.

Cmd

A single command.

Completion

Context

Encapsulates all the local state relative to a step, and is propagated to child steps.

DependencySpec

Partial representation of a Json specification of a dependency in a Cargo.toml file.

FuzzConfig

Struct representing config files for fuzzing.

FuzzableExample

Config for building an example for fuzzing.

Opt

RunCargoFuzz

RunOakExamplesOpt

Holds the options for running the example.

RunTestsOpt

RunningCmd 🔒

ServerVariantIter

An iterator over the variants of Self

SingleStatusResult

Status

A construct to keep track of the status of the execution. It only cares about the top-level steps.

StepResult

Enums

Command

Dependency

Partial representation of a dependency in a Cargo.toml file.

Scope

ServerVariant

StatusResultValue

The outcome of an individual step of execution.

Step

An execution step, which may be a single Runnable, or a collection of sub-steps.

Traits

Runnable

A task that can be run asynchronously.

Running

A task that is currently running asynchronously.

Functions

format_logs 🔒

kill_process

process_gone

run_step

Run the provided step, printing out information about the execution, and returning a set of status results from the single or multiple steps that were executed.

values_to_string 🔒