Trait xtask::task::Task

pub(crate) trait Task {
    fn get_name(&self) -> String;
    fn execute(&self) -> Result<Duration>;

    fn prepare(&mut self) -> Result<()> { ... }
    fn benchmark(&mut self, executions: u32) -> Result<Duration> { ... }
}

Required Methods

fn get_name(&self) -> String

Returns the name of the task being benchmarked

fn execute(&self) -> Result<Duration>

Executes the task to be measured and returns the time it took

Provided Methods

fn prepare(&mut self) -> Result<()>

Executes any actions required before time measurement starts By default we make sure we have a release build ready

fn benchmark(&mut self, executions: u32) -> Result<Duration>

Benchmarks the current task and returns the avarage execution time