xtask::task
Trait Task

pub(crate) trait Task {
    // Required methods
    fn get_name(&self) -> String;
    fn get_wrapped(&self, duration: Duration) -> DurationWrapper;
    fn execute(&self) -> Result<Duration>;

    // Provided methods
    fn prepare(&mut self) -> Result<()> { ... }
    fn benchmark(&mut self, executions: u32) -> Result<Duration> { ... }
}

Required Methods

fn get_name(&self) -> String

Returns the name of the task being benchmarked

fn get_wrapped(&self, duration: Duration) -> DurationWrapper

Returns a DurationWrapper with its inner type being the given argument

fn execute(&self) -> Result<Duration>

Executes the task to be measured and returns the time it took

Provided Methods

fn prepare(&mut self) -> Result<()>

Executes any actions required before time measurement starts By default we make sure we have a release build ready

fn benchmark(&mut self, executions: u32) -> Result<Duration>

Benchmarks the current task and returns the avarage execution time